refactor: Split up `IncomingPiece` code to their own functions
